+++ tests/model_executor/test_model_load_with_params.py
@@ -5,7 +5,7 @@ import os
 
 import pytest
 
-from vllm.model_executor.layers.pool import CLSPool, DispatchPooler, MeanPool
+from vllm.model_executor.layers.pooler import CLSPool, DispatchPooler, MeanPool
 from vllm.model_executor.models.bert import BertEmbeddingModel
 from vllm.model_executor.models.roberta import RobertaEmbeddingModel
 from vllm.platforms import current_platform
