+++ vllm/model_executor/layers/pooler/abstract.py
@@ -6,7 +6,7 @@ from collections.abc import Set
 import torch
 import torch.nn as nn
 
-from vllm.model_executor.layers.pool.common import PoolingParamsUpdate
+from vllm.model_executor.layers.pooler.common import PoolingParamsUpdate
 from vllm.tasks import PoolingTask
 from vllm.v1.outputs import PoolerOutput
 from vllm.v1.pool.metadata import PoolingMetadata
