+++ vllm/model_executor/layers/pooler/token/methods.py
@@ -8,7 +8,7 @@ import torch
 import torch.nn as nn
 
 from vllm.config.pooler import PoolingTypeStr
-from vllm.model_executor.layers.pool.common import PoolingParamsUpdate
+from vllm.model_executor.layers.pooler.common import PoolingParamsUpdate
 from vllm.tasks import PoolingTask
 from vllm.v1.pool.metadata import PoolingMetadata
