+++ vllm/model_executor/layers/pooler/token/poolers.py
@@ -7,11 +7,11 @@ from typing import TypeAlias
 import torch
 
 from vllm.config import PoolerConfig, get_current_vllm_config
-from vllm.model_executor.layers.pool.activations import (
+from vllm.model_executor.layers.pooler.activations import (
     PoolerActivation,
     resolve_classifier_act_fn,
 )
-from vllm.model_executor.layers.pool.common import ClassifierFn, PoolingParamsUpdate
+from vllm.model_executor.layers.pooler.common import ClassifierFn, PoolingParamsUpdate
 from vllm.tasks import PoolingTask
 from vllm.v1.pool.metadata import PoolingMetadata
