+++ vllm/model_executor/layers/pooler/tokenwise/heads.py
@@ -7,12 +7,12 @@ import torch
 import torch.nn as nn
 
 from vllm.config import get_current_vllm_config
-from vllm.model_executor.layers.pool.activations import (
+from vllm.model_executor.layers.pooler.activations import (
     PoolerActivation,
     PoolerNormalize,
     resolve_classifier_act_fn,
 )
-from vllm.model_executor.layers.pool.common import ClassifierFn
+from vllm.model_executor.layers.pooler.common import ClassifierFn
 from vllm.model_executor.models.adapters import _load_st_projector
 from vllm.pooling_params import PoolingParams
