+++ vllm/model_executor/layers/pooler/universal/poolers.py
@@ -5,7 +5,7 @@ from itertools import groupby
 
 import torch
 
-from vllm.model_executor.layers.pool.common import PoolingParamsUpdate
+from vllm.model_executor.layers.pooler.common import PoolingParamsUpdate
 from vllm.tasks import PoolingTask
 from vllm.v1.pool.metadata import PoolingMetadata
